Introduce methods for getting source numbers

+++ device/input.go
@@ -26,7 +26,7 @@ var (
 
 // NumberFromName returns the corresponding input number for the input name.
 // NOTE: The output is indexed from 1.
-func NumberFromName(device Device, input string) (uint, error) {
+func NumberFromName(device Device, input string) (uint8, error) {
 	if device > H590 {
 		return 0, errInvalidDevice
 	}
@@ -37,18 +37,18 @@ func NumberFromName(device Device, input string) (uint, error) {
 		return 0, errInvalidInput
 	}
 
-	return uint(number) + 1, nil // #nosec: Known input!
+	return uint8(number) + 1, nil // #nosec: Known input!
 }
 
 // NameFromNumber returns the corresponding input name for the input number.
 // NOTE: The input is indexed from 1.
-func NameFromNumber(device Device, number uint) (string, error) {
+func NameFromNumber(device Device, number uint8) (string, error) {
 	if device > H590 {
 		return "", errInvalidDevice
 	}
 
 	inputs := deviceInputs[device]
-	if number > uint(len(deviceInputs)) {
+	if number > uint8(len(deviceInputs)) {
 		return "", errInvalidInput
 	}
